clean up and move functions to be reusable
